Move content key type to ethportal-api
