+++ ethportal-api/Cargo.toml
@@ -13,19 +13,22 @@ authors = ["https://github.com/ethereum/trin/graphs/contributors"]
 [dependencies]
 anyhow = "1.0.68"
 bytes = "1.3.0"
+discv5 = { version = "0.2.1", features = ["serde"]}
 enr = {version= "0.7.0", features=["serde", "ed25519"]}
 ethereum-types = "0.12.1"
 eth2_ssz = "0.4.0"
 eth2_ssz_derive = "0.3.0"
 eth2_ssz_types = "0.2.1"
 jsonrpsee = {version="0.16.2", features = ["async-client", "client", "macros", "server"]}
+quickcheck = "1.0.3"
+quickcheck_macros = "1.0.0"
 reth-primitives = {git = "https://github.com/paradigmxyz/reth.git", rev = "161795945b21969faa06cd1234a0b5e3bf62f530"}
 reth-rlp = {git = "https://github.com/paradigmxyz/reth.git", rev = "161795945b21969faa06cd1234a0b5e3bf62f530"}
 serde = { version = "1.0.150", features = ["derive"] }
 serde_json = "1.0.89"
 serde-hex = "0.1.0"
 sha2 = "0.10.1"
-trin-types = { path = "../trin-types" }
+sha3 = "0.9.1"
 trin-utils = { path = "../trin-utils" }
 tokio = { version = "1.14.0", features = ["full"] }
 validator = { version = "0.13.0", features = ["derive"] }
