+++ ethportal-api/src/history.rs
@@ -1,3 +1,5 @@
+use crate::types::content_key::HistoryContentKey;
+use crate::types::enr::Enr;
 use crate::types::portal::FindNodesInfo;
 use crate::types::{
     content_item::HistoryContentItem,
@@ -7,8 +9,6 @@ use crate::types::{
     },
 };
 use jsonrpsee::{core::RpcResult, proc_macros::rpc};
-use trin_types::content_key::HistoryContentKey;
-use trin_types::enr::Enr;
 
 /// Portal History JSON-RPC endpoints
 #[rpc(client, server, namespace = "portal")]
