+++ ethportal-api/src/types/content_key.rs
