+++ ethportal-api/src/types/enr.rs
@@ -1,0 +1,3 @@
+use discv5::enr::CombinedKey;
+
+pub type Enr = discv5::enr::Enr<CombinedKey>;
