+++ ethportal-api/src/types/mod.rs
@@ -1,4 +1,5 @@
 pub mod content_item;
+pub mod content_key;
 pub mod discv5;
+pub mod enr;
 pub mod portal;
-pub mod request;
