+++ ethportal-api/src/types/portal.rs
@@ -1,8 +1,8 @@
 use crate::types::content_item::HistoryContentItem;
+use crate::types::content_key::HistoryContentKey;
+use crate::types::enr::Enr;
 use serde::{Deserialize, Serialize};
 use ssz_types::{typenum, BitList};
-use trin_types::content_key::HistoryContentKey;
-use trin_types::enr::Enr;
 
 pub type DataRadius = ethereum_types::U256;
 pub type Distance = ethereum_types::U256;
