+++ rpc/src/server.rs
@@ -1,12 +1,12 @@
 use crate::jsonrpsee::server::{ServerBuilder as HttpServerBuilder, ServerHandle};
 use crate::{Discv5Api, HistoryNetworkApi, Web3Api};
-use ethportal_api::types::request::HistoryJsonRpcRequest;
 use ethportal_api::{Discv5ApiServer, HistoryNetworkApiServer, Web3ApiServer};
 use reth_ipc::server::Builder as IpcServerBuilder;
 use std::net::SocketAddr;
 use std::sync::Arc;
 use tokio::sync::mpsc;
 use trin_core::portalnet::discovery::Discovery;
+use trin_types::request::HistoryJsonRpcRequest;
 
 use url::Url;
