+++ src/bin/purge_db.rs
@@ -10,11 +10,11 @@ use tracing::{info, warn};
 use ethportal_api::types::content_item::{
     BlockBody, ContentItem, EpochAccumulator, HeaderWithProof,
 };
+use ethportal_api::HistoryContentKey;
 use ethportal_api::Receipt;
 use trin_core::portalnet::storage::{PortalStorage, PortalStorageConfig};
 use trin_core::portalnet::types::messages::ProtocolId;
 use trin_core::utils::db::get_data_dir;
-use trin_types::content_key::HistoryContentKey;
 
 ///
 /// This script will iterate through all content id / key pairs in rocksd & meta db.
