+++ trin-cli/Cargo.toml
@@ -14,6 +14,7 @@ authors = ["https://github.com/ethereum/trin/graphs/contributors"]
 anyhow = "1.0.68"
 base64 = "0.13.0"
 ethereum-types = "0.12.1"
+ethportal-api = { path = "../ethportal-api" }
 jsonrpc = "0.12.0"
 nanotemplate = "0.3.0"
 serde = "1.0.150"
