+++ trin-state/src/lib.rs
@@ -11,13 +11,13 @@ use tokio::{
 use tracing::info;
 
 use crate::{events::StateEvents, jsonrpc::StateRequestHandler};
-use ethportal_api::types::request::StateJsonRpcRequest;
 use trin_core::{
     portalnet::{
         discovery::Discovery, storage::PortalStorageConfig, types::messages::PortalnetConfig,
     },
     utp::stream::{UtpListenerEvent, UtpListenerRequest},
 };
+use trin_types::request::StateJsonRpcRequest;
 use trin_validation::oracle::HeaderOracle;
 
 pub mod events;
