+++ trin-state/src/network.rs
@@ -5,6 +5,7 @@ use eth_trie::EthTrie;
 use parking_lot::RwLock as PLRwLock;
 use tokio::sync::{mpsc::UnboundedSender, RwLock};
 
+use ethportal_api::StateContentKey;
 use trin_core::{
     portalnet::{
         discovery::Discovery,
@@ -14,7 +15,6 @@ use trin_core::{
     },
     utp::stream::UtpListenerRequest,
 };
-use trin_types::content_key::StateContentKey;
 use trin_types::distance::XorMetric;
 use trin_validation::oracle::HeaderOracle;
