+++ trin-state/src/validation.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use tokio::sync::RwLock;
 
-use trin_types::content_key::StateContentKey;
+use ethportal_api::StateContentKey;
 use trin_validation::{oracle::HeaderOracle, validator::Validator};
 
 pub struct StateValidator {
