+++ trin-types/Cargo.toml
@@ -18,6 +18,7 @@ clap = "2.33.3"
 discv5 = { version = "0.2.1", features = ["serde"]}
 eth_trie = "0.1.0"
 ethereum-types = "0.12.1"
+ethportal-api = { path = "../ethportal-api" }
 eth2_ssz = "0.4.0"
 eth2_ssz_derive = "0.3.0"
 eth2_ssz_types = "0.2.1"
@@ -31,6 +32,7 @@ serde_json = "1.0.89"
 sha2 = "0.10.1"
 sha3 = "0.9.1"
 structopt = "0.3.26"
+tokio = { version = "1.14.0", features = ["full"] }
 trin-utils = { path = "../trin-utils" }
 ureq = { version = "2.5.0", features = ["json"] }
 url = "2.3.1"
