+++ trin-types/src/lib.rs
@@ -1,8 +1,8 @@
 pub mod bytes;
 pub mod cli;
-pub mod content_key;
 pub mod distance;
 pub mod enr;
 pub mod execution;
 pub mod jsonrpc;
 pub mod provider;
+pub mod request;
