+++ trin-validation/src/accumulator.rs
@@ -16,9 +16,9 @@ use crate::{
     merkle::proof::{verify_merkle_proof, MerkleTree},
 };
 use ethportal_api::endpoints::HistoryEndpoint;
-use ethportal_api::types::request::HistoryJsonRpcRequest;
-use trin_types::content_key::{EpochAccumulatorKey, HistoryContentKey};
+use ethportal_api::types::content_key::{EpochAccumulatorKey, HistoryContentKey};
 use trin_types::execution::header::{BlockHeaderProof, Header, HeaderWithProof};
+use trin_types::request::HistoryJsonRpcRequest;
 use trin_utils::bytes::hex_decode;
 
 /// SSZ List[Hash256, max_length = MAX_HISTORICAL_EPOCHS]
